requirements.txt
django-cms>=3.8,<3.9
djangocms-admin-style>=2.0,<3.0
django-treebeard>=4.0,<5.0

djangocms-text-ckeditor>=4.0,<5.0
djangocms-link>=3.0,<4.0
djangocms-icon>=2.0,<3.0
djangocms-style>=3.0,<4.0
djangocms-googlemap>=2.0,<3.0
djangocms-video>=3.0,<4.0
djangocms-file>=3.0,<4.0
djangocms-picture>=3.0,<4.0
djangocms-bootstrap4>=2.0,<3.0
easy_thumbnails
django-filer>=1.3
Django<3.2
django-classy-tags>=2.0
django-sekizai>=2.0
django-mptt>0.9
html5lib>=1.0.1
Pillow>=3.0
six
pytz